address TODO

+++ include/kernel/syscall.h
@@ -51,10 +51,9 @@ static inline void syscall_args_set_return_ptr(jinue_syscall_args_t *args, void 
 	syscall_args_set_return_uintptr(args, (uintptr_t)retval);
 }
 
-/* TODO do we need to do something to handle int vs (u)intptr_t for the error number? */
-static inline void syscall_args_set_error(jinue_syscall_args_t *args, uintptr_t error) {
+static inline void syscall_args_set_error(jinue_syscall_args_t *args, int error) {
 	args->arg0	= (uintptr_t)-1;
-	args->arg1	= error;
+	args->arg1	= (uintptr_t)error;
 	args->arg2	= 0;
 	args->arg3	= 0;
 }
